% 
% TODO: Write Documentation 
% 
\name{fetch}

\alias{fetch}

\title{Retrieving Weather data from GeoClimate Database System}

\description{
	% TODO: Write description
}

\usage{
		% TODO: Write usage e.g. fun(arg1, arg2, …)
		fetch.daily(xy, sv, ...)
		fetch.monthly(xy, sv, ...)
}

\arguments{
		% TODO: Write arguments
}

\details{
		% TODO: Write details
}

\value{
		% TODO: Write value
}

\references{
		% TODO: Write references
}

\note{
		% TODO: Write note
}

\author{Jorrel Khalil S. Aunario}

\seealso{
		% TODO: Write see also
}

\examples{
		% TODO: Write examples
}

\keyword{
		% TODO: Write key
}
