% 
% TODO: Write Documentation 
% 
\name{Fortran Simulation Environment }

\alias{
		% TODO: Write alias
}

\title{Reading and Writing Fortran Simulation Evironment (FSE) Weather files}

\description{
	% TODO: Write description
}

\usage{
		% TODO: Write usage e.g. fun(arg1, arg2, …)
}

\arguments{
		% TODO: Write arguments
}

\details{
		% TODO: Write details
}

\value{
		% TODO: Write value
}

\references{
		% TODO: Write references
}

\note{
		% TODO: Write note
}

\author{Jorrel Khalil S. Aunario}

\seealso{
		% TODO: Write see also
}

\examples{
		% TODO: Write examples
}

\keyword{
		% TODO: Write key
}
