\name{weather-package}
\alias{weather-package}
\alias{weather}
\docType{package}
\title{
What the package does (short line)
	Weather data manipulation
}
\description{
More about what it does (maybe more than one line)
	Weather data manipulation. Particularly in the context of ecological modeling, such as crop growth simulation modeling.
}
\details{
\tabular{ll}{
Package: \tab weather\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-10-17\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Robert J. Hijmans
Maintainer: <r.hijmans@gmail.com>
}

\keyword{ package }
\keyword{ spatial }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
